Force new users to change password upon initial portal login.
All new students have an assigned user name and the same password. Users logging into the system with the original default password must change the password.

AccountController - Changed to user user name instead of email, and to redirect the user to the change password if they logged in with the default password. The default code was setting the model.code to null in the Reset Password method; so, I added code to get the token to pass for the code parameter.

ManageController - changed to redirect the user to the portal main page after successfully changing a password.

AccountViewModel - changed to use UserName instead of Email.